# apiserver/swagger_server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a Recommandation Micro Service"
  version: "1.0.0"
  title: "Recommandation"
basePath: "/"
tags:
- name: "api"
  description: "Microservice API"
schemes:
- "https"
- "http"
paths:
  /getRestaurants:
    post:
      tags:
      - "api"
      summary: "return restaurant list by keyword"
      description: ""
      operationId: "get_restaurants"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Keyword"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getCategory:
    post:
      tags:
      - "api"
      summary: "return category list"
      description: ""
      operationId: "get_category"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Nothing (Empty data)"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getFood:
    post:
      tags:
      - "api"
      summary: "return food list by category"
      description: ""
      operationId: "get_food"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Category"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getRestaurant:
    post:
      tags:
      - "api"
      summary: "return selected restaurant detail"
      description: ""
      operationId: "get_restaurant"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "restaurant"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getRestaurantHistory:
    post:
      tags:
      - "api"
      summary: "return restaurant list that the target user selected"
      description: ""
      operationId: "get_restaurant_history"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User Information"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getFoodHistory:
    post:
      tags:
      - "api"
      summary: "return food list that the target user selected"
      description: ""
      operationId: "get_food_history"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User Information"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getRestaurantbyDefault:
    post:
      tags:
      - "api"
      summary: "return restaurant list by user info"
      description: ""
      operationId: "get_restaurantby_default"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User Info"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getRestaurantbyKeyword:
    post:
      tags:
      - "api"
      summary: "return restaurant list by keyword and user info"
      description: ""
      operationId: "get_restaurantby_keyword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Keyword and User Info"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getFoodbyCategory:
    post:
      tags:
      - "api"
      summary: "return food list by category and restaurant list by user info"
      description: ""
      operationId: "get_foodby_category"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Category and User Info"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getRestaurantbyFood:
    post:
      tags:
      - "api"
      summary: "return restaurant list by food and user"
      description: ""
      operationId: "get_restaurantby_food"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Food and User Info"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
  /getUserInfo:
    post:
      tags:
      - "api"
      summary: "return user information. for testing purpose, currently only one user(guest)"
      description: ""
      operationId: "get_user_info"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Nothing (Empty data)"
        required: true
        schema:
          $ref: "#/definitions/data"
      responses:
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "swagger_server.controllers.api_controller"
definitions:
  data:
    type: "object"
    required:
    - "data"
    properties:
      data:
        type: "object"
        properties: {}